FEATURE 25: REPORTES DE ESTADISTICA ECONOMICOS Y FIX VARIOS

+++ src/services/TableService.ts
@@ -1,3 +1,6 @@
+
+import { Order } from "../models/Order";
+import { Session } from "../models/Session";
 import { Table } from "../models/Table";
 
 export class TableService {
@@ -22,6 +25,27 @@ export class TableService {
 
     public async deleteTable(id: string) {
         const table = new Table({ tableId: id, tableNumber: 0, status: "Disponible" });
+
+        //Verificar si la mesa tiene sessiones activas
+        const sessions = await Session.getAllSessions();
+
+        const activeSession = sessions.find(session => session.table.tableId === id && session.status === 'Activa');
+
+        if (activeSession) {
+            throw new Error('No se puede eliminar la mesa porque tiene una sesión activa');
+        }
+
+        //Verificar si la mesa tiene sessiones o pedidos asociados
+
+        const orders = await Order.getAll();
+
+        const associatedOrders = orders.find(order => order.table.tableId === id);
+        const associatedSessions = sessions.find(session => session.table.tableId === id);
+
+        if (associatedOrders || associatedSessions) {
+            throw new Error('No se puede eliminar la mesa porque tiene pedidos asociados');
+        }
+        
         return await table.delete();
     }
 }
